Replaced Activities class with a parse_data function
Was only using init for the Activities class and storing a defaultdict(Activity) inside it. Instead of using a constructor I now just have a factory method to create this defaultdict(Activity) and return it.

Next, I'll update this function (parse_data) to actually return the defaultdict as well as a list of behaviours (along with ground truth)